refactor: extract RecordingSessionStopReadiness.swift from RecordingSessionOutcomes.swift
Splits the `RecordingSessionStopReadiness` enum out of
RecordingSessionOutcomes.swift.

Byte-identical move of lines 12-17. Both files import Foundation.

RecordingSessionOutcomes.swift: 22 → 15 lines. Tests: 667.
