+++ Sources/BugNarrator/Services/RecordingSessionOutcomes.swift
@@ -9,13 +9,6 @@ enum RecordingSessionStartOutcome {
     case failure(Error)
 }
 
-enum RecordingSessionStopReadiness {
-    case ready(RecordingSessionDraft)
-    case transitionInProgress
-    case noActiveRecording
-    case missingSessionMetadata
-}
-
 enum RecordingSessionCancelOutcome {
     case cancelled(RecordingSessionDraft?)
     case transitionInProgress
