+++ Sources/BugNarrator/Services/RecordingSessionStopReadiness.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum RecordingSessionStopReadiness {
+    case ready(RecordingSessionDraft)
+    case transitionInProgress
+    case noActiveRecording
+    case missingSessionMetadata
+}
